Bugfixes. Update version and changelog.

+++ robo_bot_cli/commands/connect.py
@@ -1,25 +1,57 @@
 import math
 import os
-from os.path import join, dirname
+from os.path import dirname, isfile, join
 from shutil import copyfile
 
-import pkg_resources
-
 import click
-from questionary import prompt, Choice
-from robo_bot_cli.config.bot_manifest import BotManifest
+import pkg_resources
+from questionary import Choice, prompt
 from robo_ai.model.assistant.assistant import Assistant
-from robo_ai.model.assistant.assistant_list_response import AssistantListResponse
-from robo_bot_cli.util.cli import loading_indicator, print_message, print_success
-from robo_bot_cli.util.robo import get_robo_client, validate_robo_session, validate_bot, get_supported_base_versions
+from robo_ai.model.assistant.assistant_list_response import (
+    AssistantListResponse,
+)
+
+from robo_bot_cli.config.bot_manifest import BotManifest
 from robo_bot_cli.config.tool_settings import ToolSettings
+from robo_bot_cli.util.cli import (
+    loading_indicator,
+    print_message,
+    print_success,
+)
+from robo_bot_cli.util.robo import (
+    get_robo_client,
+    get_supported_base_versions,
+    validate_bot,
+    validate_robo_session,
+)
 
 
-@click.command(name='connect', help='Connect a local bot to a ROBO.AI server bot instance.')
-@click.argument('language', nargs=-1,)
-@click.option('--bot-uuid', default=None, type=str, help='The bot UUID to assign to the bot implementation.')
-@click.option('--target-dir', default=None, type=str, help='The target directory where the bot will be setup.')
-def command(language: tuple, bot_uuid: str, target_dir: str, base_version: str = 'rasa-1.10.0'):
+@click.command(
+    name="connect",
+    help="Connect a local bot to a ROBO.AI server bot instance.",
+)
+@click.argument(
+    "language",
+    nargs=-1,
+)
+@click.option(
+    "--bot-uuid",
+    default=None,
+    type=str,
+    help="The bot UUID to assign to the bot implementation.",
+)
+@click.option(
+    "--target-dir",
+    default=None,
+    type=str,
+    help="The target directory where the bot will be setup.",
+)
+def command(
+    language: tuple,
+    bot_uuid: str,
+    target_dir: str,
+    base_version: str = "rasa-1.10.0",
+):
     """
     Connect a local bot to a ROBO.AI server bot instance.
     This instance must be already created in the ROBO.AI platform.
@@ -38,17 +70,23 @@ def command(language: tuple, bot_uuid: str, target_dir: str, base_version: str =
     validate_bot(bot_uuid)
 
     if len(language) == 0:
-        bot_dir = bot_ignore_dir = os.path.abspath(target_dir) \
-                                   if target_dir else create_implementation_directory(bot_uuid)
+        bot_dir = bot_ignore_dir = (
+            os.path.abspath(target_dir)
+            if target_dir
+            else create_implementation_directory(bot_uuid)
+        )
     elif len(language) == 1:
-        bot_dir = os.path.abspath(join(target_dir, 'languages', language[0])) \
-                  if target_dir else create_implementation_directory(bot_uuid)  # TODO check what this does
+        bot_dir = (
+            os.path.abspath(join(target_dir, "languages", language[0]))
+            if target_dir
+            else create_implementation_directory(bot_uuid)
+        )  # TODO check what this does
         bot_ignore_dir = join(dirname(dirname(bot_dir)))
     else:
-        print_message('Please select only one bot to connect to.')
+        print_message("Please select only one bot to connect to.")
         exit(0)
 
-    print_message('Bot target directory: {0}'.format(bot_dir))
+    print_message("Bot target directory: {0}".format(bot_dir))
 
     create_bot_manifest(bot_dir, bot_uuid, base_version)
     create_bot_ignore(bot_ignore_dir)
@@ -82,8 +120,8 @@ def get_bot_choice(bot: Assistant) -> dict:
         dict: dictionary with assistant's name and ID.
     """
     return {
-        'name': bot.name,
-        'value': bot.uuid,
+        "name": bot.name,
+        "value": bot.uuid,
     }
 
 
@@ -94,9 +132,9 @@ def get_bot_uuid() -> str:
     Returns:
         bot_uuid (str): ID of the selected bot.
     """
-    NEXT_PAGE = '__NEXT_PAGE__'
-    PREV_PAGE = '__PREV_PAGE__'
-    NONE = '__NONE__'
+    NEXT_PAGE = "__NEXT_PAGE__"
+    PREV_PAGE = "__PREV_PAGE__"
+    NONE = "__NONE__"
     META_RESPONSES = [NEXT_PAGE, PREV_PAGE, NONE]
 
     bot_uuid = NONE
@@ -113,31 +151,32 @@ def get_bot_uuid() -> str:
         bot_choices = list(map(get_bot_choice, bots.content))
         page_count = math.ceil(bots.totalElements / bots.size)
         if page < page_count - 1:
-            bot_choices.append({
-                'name': '> Next page...',
-                'value': NEXT_PAGE
-            })
+            bot_choices.append({"name": "> Next page...", "value": NEXT_PAGE})
 
         if page > 0:
-            bot_choices.append({
-                'name': '> Previous page...',
-                'value': PREV_PAGE
-            })
+            bot_choices.append(
+                {"name": "> Previous page...", "value": PREV_PAGE}
+            )
 
         questions = [
             {
-                'type': 'list',
-                'name': 'bot_id',
-                'message': 'Please select the bot you would like to implement:',
-                'choices': [Choice(title=bot['name'], value=bot['value'])
-                            if (bot['value'] == NEXT_PAGE or bot['value'] == PREV_PAGE)
-                            else Choice(title=str(bot['name'] + ' ('+bot['value']+')'), value=bot['value'])
-                            for bot in bot_choices],
+                "type": "list",
+                "name": "bot_id",
+                "message": "Please select the bot you would like to implement:",
+                "choices": [
+                    Choice(title=bot["name"], value=bot["value"])
+                    if (bot["value"] == NEXT_PAGE or bot["value"] == PREV_PAGE)
+                    else Choice(
+                        title=str(bot["name"] + " (" + bot["value"] + ")"),
+                        value=bot["value"],
+                    )
+                    for bot in bot_choices
+                ],
             },
         ]
 
         answers = prompt(questions)
-        bot_uuid = answers['bot_id']
+        bot_uuid = answers["bot_id"]
 
     return bot_uuid
 
@@ -152,28 +191,28 @@ def get_base_version() -> str:
     with loading_indicator("Loading base version list..."):
         versions = get_supported_base_versions()
 
-    version_choices = [{
-        'value': base_version['version'],
-        'name': base_version['label']
-    } for base_version in versions]
+    version_choices = [
+        {"value": base_version["version"], "name": base_version["label"]}
+        for base_version in versions
+    ]
 
     questions = [
         {
-            'type': 'list',
-            'name': 'base_version',
-            'message': 'Please select the bot runtime version:',
-            'choices': version_choices
+            "type": "list",
+            "name": "base_version",
+            "message": "Please select the bot runtime version:",
+            "choices": version_choices,
         },
     ]
     answers = prompt(questions)
-    base_version = answers['base_version']
+    base_version = answers["base_version"]
 
     return base_version
 
 
 def create_implementation_directory(bot_uuid: str) -> str:
     cwd = os.getcwd()
-    bot_dir = cwd + '/' + bot_uuid
+    bot_dir = cwd + "/" + bot_uuid
     os.mkdir(bot_dir)
     return bot_dir
 
@@ -186,11 +225,14 @@ def create_bot_manifest(bot_dir: str, bot_uuid: str, base_version: str):
 
 
 def create_bot_ignore(bot_ignore_dir: str):
-    copyfile(get_bot_ignore_path(), join(bot_ignore_dir, '.botignore'))
+    if not isfile(join(bot_ignore_dir, ".botignore")):
+        copyfile(get_bot_ignore_path(), join(bot_ignore_dir, ".botignore"))
 
 
 def get_bot_ignore_path() -> str:
-    return pkg_resources.resource_filename(__name__, "../initial_structure/initial_project/.botignore")
+    return pkg_resources.resource_filename(
+        __name__, "../initial_structure/initial_project/.botignore"
+    )
 
 
 if __name__ == "__main__":
